Replace timer check with RegistryMonitor
Instead of periodically querying the registry, we now set up a
RegistryMonitor to trigger an event any time a value within the set
that we care about changes.
This also meant an overhaul of the GUI because its main purpose was no
longer necessary. It is now more informational, rather than a settings
dialog.

Also added functionality to retrieve the paths of the current wallpapers.
This information is displayed in the GUI and the files can be opened
directly from here.