Fix threading & stretched wallpaper issues
If a single wallpaper is stretched across multiple monitors, only one
TranscodedImageCache will be set, so we need to deal with that scenario.

Also added code to make setting GUI values possible from CheckWallpaper,
by handing it off to the thread that created the GUI to call the
function.

Also changed line-wrapping to 100 chars, instead of 80.